refactor: clean up wrapper entrypoint
